+++ Final_Project/Final_Project/Party.h
@@ -17,7 +17,7 @@ private:
 	vector<TcpSocket*> _sockets;//socket for all TCP connection. sockets[0] is a UDP soket
 	vector<pair<long, long>> _shares;//index of vector is the id of input's party
 public:
-	Party(int myID,long input);
+	Party(short myID,long input);
 	void connectToAllParties(string IPs[NUM_OF_PARTIES]);
 	void broadcast(char * msg)const;
 	~Party();
