+++ Final_Project/Final_Project/TcpSocket.cpp
@@ -3,7 +3,7 @@
 
 using std::cerr;
 WSAInitializer TcpSocket::_WSAinit;
-TcpSocket::TcpSocket(int socket, int port) :_port(port)
+TcpSocket::TcpSocket(int socket, unsigned short port):_port(port)
 {
 	_socket = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
 
@@ -12,34 +12,34 @@ TcpSocket::TcpSocket(int socket, int port) :_port(port)
 
 }
 
-	TcpSocket :: ~TcpSocket()
-	{
-		if(isValid())
-			this->close();
+TcpSocket :: ~TcpSocket()
+{
+	if(isValid())
+		this->close();
 
-	}
+}
 
-	int TcpSocket::writeBuffer(const void* buffer, long bufferSize, int flags)
-	{
-		if (!isValid()) return -1;
-		return send(_socket, (const char*)buffer, sizeof(buffer), flags);
-	}
+int TcpSocket::writeBuffer(const void* buffer, long bufferSize, int flags)
+{
+	if (!isValid()) return -1;
+	return send(_socket, (const char*)buffer, sizeof(buffer), flags);
+}
 
-	int TcpSocket::readBuffer(void* buffer, long bufferSize, int flags)
-	{
-		if (!isValid()) return -1;
-		return recv(_socket, (char*)buffer, bufferSize, flags);
-	}
-	void TcpSocket::close()
-	{
-		if (!isValid()) return;
+int TcpSocket::readBuffer(void* buffer, long bufferSize, int flags)
+{
+	if (!isValid()) return -1;
+	return recv(_socket, (char*)buffer, bufferSize, flags);
+}
+void TcpSocket::close()
+{
+	if (!isValid()) return;
 
-		::closesocket(_socket);
+	::closesocket(_socket);
 
-	}
-	bool TcpSocket::isValid()const {
-		return this->_socket != -1;
-	}
-	int TcpSocket::socketFd()const { 
-		return isValid() ? _socket : -1; 
-	}
+}
+bool TcpSocket::isValid()const {
+	return this->_socket != -1;
+}
+int TcpSocket::socketFd()const { 
+	return isValid() ? _socket : -1; 
+}
